Add some folders to organise the source a bit better
