+++ src/bankHolidays.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const bankHolidayJson = require('./bank-holidays.json');
+const bankHolidayJson = require('../data/bank-holidays.json');
 
 const bankHolidays = bankHolidayJson['england-and-wales'].events.map(holiday => holiday.date);
