+++ src/tally.js
